requirements.txt
pdfplumber
python-dotenv
openai
langchain-openai
langchain-core
langchain-community
fastapi
uvicorn
Pillow  
pydantic
faiss-cpu 
